pyproject.toml
[project]
name = "aletheia"
version = "0.1.0"
description = "AI-powered troubleshooting tool for SREs and system administrators"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Aletheia Contributors"}
]
keywords = ["sre", "troubleshooting", "observability", "ai", "llm", "incident-response"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "pycryptodome>=3.19",
    "rich>=13.7",
    "requests>=2.31",
    "gitpython>=3.1",
    "cryptography>=41.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "platformdirs>=4.0.0",
    "dateparser>=1.2.2",
    "agent-framework>=1.0.0b260107",
    "agent-framework-bedrock>=1.0.0b260107",
    "azure-ai-agents==1.2.0b5",
    "jinja2>=3.1.6",
    "scapy>=2.7.0rc1",
    "pylint>=4.0.3",
    "boto3>=1.26.0",
    "docker>=7.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.9",
    "sse-starlette>=2.0.0",
    "chromadb>=1.4.0",
    "prompt-toolkit>=3.0.43",
    "azure-identity>=1.25.1",
    "azure-mgmt-resource>=24.0.0",
    "azure-mgmt-compute>=37.1.0",
    "azure-mgmt-storage>=24.0.0",
    "azure-mgmt-keyvault>=13.0.0",
    "azure-keyvault-keys>=4.11.0",
    "azure-keyvault-secrets>=4.10.0",
    "azure-mgmt-cdn>=13.1.1",
    "azure-monitor-query>=2.0.0",
    "python-telegram-bot[all]>=21.0.0",
    "websockets>=12.0",
    "black>=25.12.0",
    "structlog>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.7.0",
    "ruff>=0.0.286",
    "mypy>=1.5.0",
    "types-pyyaml>=6.0",
    "types-requests>=2.31",
]
ast = [
    "tree-sitter>=0.20",
]

[project.scripts]
aletheia = "aletheia.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aletheia"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=aletheia",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]

[tool.coverage.run]
branch = true
source = ["aletheia"]
omit = [
    "tests/*",
    "**/__pycache__/*",
    "**/.*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 88
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests/*" = ["S101"]      # assert allowed in tests

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true

[[tool.mypy.overrides]]
module = [
    "semantic_kernel.*",
    "gitpython.*",
]
ignore_missing_imports = true
